Bugfix, updates and, improvements; - Remove hardcode configuration - Improve error handling - Improve message of error with documentation - Added README.md - Added gitignore

+++ api.py
@@ -3,65 +3,74 @@ Script to interact with VMware Fusion API to control VMs.
 """
 # -*- coding: utf-8 -*-
 
+import os
 import json
 import subprocess
 import sys
 import requests
 
-URL = "http://localhost:8697/api/vms"
-BASIC_AUTH = 'Basic ZmxvcmVudGlubzpGPWU1YWNkMTgwOA=='
+VMREST_URL = os.environ.get('VMREST_URL')
+VMREST_AUTH = os.environ.get('VMREST_AUTH')
 PAYLOAD = {}
 HEADERS = {
     'Accept': 'application/vnd.vmware.vmw.rest-v1+json',
-    'Authorization': BASIC_AUTH
+    'Authorization': VMREST_AUTH
 }
 
 
 def enable_fusion_api(function):
     """This function is used to enable or disable the Fusion API"""
-    if function == "enable":
-        cmd_ena_api = "nohup vmrest > /tmp/vmrest.out 2>&1 &"
-        subprocess.call(cmd_ena_api, shell=True)
-        msg = "VMware Fusion API enabled"
+    try:
+        if VMREST_URL:
+            if function == "enable":
+                try:
+                    cmd_ena_api = "nohup vmrest > /tmp/vmrest.out 2>&1 &"
+                    subprocess.call(cmd_ena_api, shell=True)
+                    msg = "VMware Fusion API enabled"
+                except ValueError:
+                    msg = "VMware Fusion API already enabled"
 
-    if function == "disable":
-        try:
-            cmd_get_pid = "ps aux |grep vmrest |grep -v grep |awk '{ print $2 }'"
-            with subprocess.Popen(cmd_get_pid, shell=True, stdout=subprocess.PIPE,
-                                  stderr=subprocess.STDOUT) as proc:
-                result = int(proc.communicate()[0].decode('utf-8'))
-                cmd_kill_pid = "kill -9 " + str(result)
-            with subprocess.Popen(cmd_kill_pid, shell=True,
-                                  stdout=subprocess.PIPE,
-                                  stderr=subprocess.STDOUT) as process:
-                process.wait()
-                msg = "VMware Fusion API disabled"
-        except ValueError:
-            msg = "VMware Fusion API already disabled"
+            if function == "disable":
+                try:
+                    cmd_get_pid = "ps aux |grep vmrest |grep -v grep |awk '{ print $2 }'"
+                    with subprocess.Popen(cmd_get_pid, shell=True, stdout=subprocess.PIPE,
+                                          stderr=subprocess.STDOUT) as proc:
+                        result = int(proc.communicate()[0].decode('utf-8'))
+                        cmd_kill_pid = "kill -9 " + str(result)
+                    with subprocess.Popen(cmd_kill_pid, shell=True,
+                                          stdout=subprocess.PIPE,
+                                          stderr=subprocess.STDOUT) as process:
+                        process.wait()
+                        msg = "VMware Fusion API disabled"
+                except ValueError:
+                    msg = "VMware Fusion API already disabled"
+        else:
+            msg = "VMREST_URL not set, please set it in the environment variable: \
+                   \nEx. export VMREST_URL=http://127.0.0.1:8697/api/vms"
+    except ValueError:
+        msg = "VMware Fusion API is not enabled"
 
     return print(msg)
 
 
 def get_vm_list():
     """This function is used to get the list of VMs"""
-    try:
-        response = requests.request("GET", URL, headers=HEADERS, data=PAYLOAD)
-        return print(json.dumps(response.json(), indent=4, sort_keys=True))
-    except requests.exceptions.RequestException:
-        return print("VMware Fusion API is not enabled")
+    response = requests.request(
+        "GET", VMREST_URL, headers=HEADERS, data=PAYLOAD)
+    return print(json.dumps(response.json(), indent=4, sort_keys=True))
 
 
 def get_vm_info(vm_id):
     """ This function is used to get the info of a VM """
     response = requests.request(
-        "GET", URL + "/" + vm_id, headers=HEADERS, data=PAYLOAD)
+        "GET", VMREST_URL + "/" + vm_id, headers=HEADERS, data=PAYLOAD)
     return response.json()
 
 
 def get_vm_power_state(vm_id):
     """This function is used to get the power state of a VM"""
     response = requests.request(
-        "GET", URL + "/" + vm_id + "/power", headers=HEADERS, data=PAYLOAD)
+        "GET", VMREST_URL + "/" + vm_id + "/power", headers=HEADERS, data=PAYLOAD)
     return response.json()
 
 
@@ -71,9 +80,6 @@ if __name__ == '__main__':
             enable_fusion_api(sys.argv[2])
         if "--get-vm-list" in sys.argv[1]:
             get_vm_list()
-        else:
-            print("Usage: vmware_fusion.py --api [enable|disable]")
-            print("Usage: vmware_fusion.py --get-vm-list")
     except ConnectionRefusedError:
         sys.exit("\nThe host is not responding or don't exist\n")
     except requests.exceptions.ConnectionError:
